+++ .eleventy.js
@@ -1,0 +1,11 @@
+module.exports = function (eleventyConfig) {
+  return {
+    dir: {
+      data: 'data',
+      includes: 'includes',
+      input: 'src/pshry.com',
+      layouts: 'layouts',
+      output: 'build/pshry.com'
+    }
+  }
+}